+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Binders/Renderers/Enabled/Mono/RendererEnabledEnumGroupMonoBinder.cs
@@ -1,0 +1,20 @@
+using UnityEngine;
+
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM.StarterKit
+{
+    /// <summary>
+    /// <see cref="EnumGroupMonoBinder{Renderer, Boolean}"/> that sets <see cref="Renderer.enabled"/>
+    /// on each element in the group based on the bound enum ViewModel value.
+    /// </summary>
+    [AddBinderContextMenu(typeof(Renderer), serializePropertyNames: "m_Enabled", SubPath = "EnumGroup")]
+    [AddComponentMenu("Aspid/MVVM/Binders/Renderer/Renderer Binder – Enabled EnumGroup")]
+    public sealed class RendererEnabledEnumGroupMonoBinder : EnumGroupMonoBinder<Renderer, bool>
+    {
+        /// <summary>
+        /// Called when the bound enum resolves to a value for the specified element.
+        /// </summary>
+        protected override void SetValue(Renderer element, bool value) =>
+            element.enabled = value;
+    }
+}